requirements.txt
anyio==4.6.0
APScheduler==3.10.4
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
python-telegram-bot==21.6
pytz==2024.2
six==1.16.0
sniffio==1.3.1
tzdata==2024.2
tzlocal==5.2
